api/requirements.txt
fastapi==0.111.1
PyYAML==6.0.2
python-dotenv==1.0.1
requests==2.32.3
uvicorn[standard]==0.30.4
